func-tests: move common function to utilities and add more asserts
