+++ functional-tests/fn_btcio_inscriber.py
@@ -5,7 +5,7 @@ from bitcoinlib.services.bitcoind import BitcoindClient
 
 import testenv
 from constants import SEQ_PUBLISH_BATCH_INTERVAL_SECS
-from utils import generate_n_blocks, wait_until
+from utils import generate_n_blocks, submit_da_blob, wait_until
 
 
 @flexitest.register
@@ -33,20 +33,10 @@ class L1WriterTest(testenv.StrataTester):
 
         # Submit blob
         blobdata = "2c4253d512da5bb4223f10e8e6017ede69cc63d6e6126916f4b68a1830b7f805"
-        _ = seqrpc.strataadmin_submitDABlob(blobdata)
-
-        # Allow some time for sequencer to publish blob
-        time.sleep(SEQ_PUBLISH_BATCH_INTERVAL_SECS)
-
-        l1_status = seqrpc.strata_l1status()
-        txid = l1_status["last_published_txid"]
-
+        tx = submit_da_blob(btcrpc, seqrpc, blobdata)
         # Calculate scriptbpubkey for sequencer address
         addrdata = btcrpc.proxy.validateaddress(seqaddr)
         scriptpubkey = addrdata["scriptPubKey"]
-
-        # Check if txn is present in mempool/blockchain and is spent to sequencer address
-        tx = btcrpc.gettransaction(txid)
 
         # NOTE: could have just compared address
         # but bitcoinlib is somehow giving bc1* addr even though network is regtest
